Fix Kodi plugin.video.youtube... again!
This commit bumps the `plugin.video.youtube` Kodi plugin to the latest
version (6.3.1).

Because the repository we were using is only providing the latest
version of the plugin, our build was broken (due to a 404 error) every
time a new version of the plugin was released. To avoid this issue in
the future, this commit also changes the download source to fetch the
plugin directly from the GitHub repository, rather than from the Kodi
addons repository (the GitHub repository is a lot more likely to keep
providing all released versions).