Infer ARCH from directory name
The script that compiles the project in a Docker container requires
developers to always have $ARCH defined in their env (or to define it
everytime they invoke the script).

Since the build instructions in both the README and the wiki suggest to
clone the git repository into a directory named `recalbox-${ARCH}`, this
commit tries to infer $ARCH from the directory name we are currently
working in.

It means that if I'm working in a directory named `recalbox-rpi3`, I
don't need to specify `$ARCH` each time I run
`script/linux/recaldocker.sh` anymore: it will be infered to `rpi3`
automatically.

Only architectures with an existing defconfig are recognised, though, to
avoid mistaking `recalbox-my_fork`, `recalbox-latest` as
architectures.