Fix core renames migration regexp
In 1479c5b, we introduced a shell script that migrates `recalbox.conf`
on boot to use new core names introduced in bf151a1c.

Unfortunately, the regexp we use was not strict enough. In particular,
it was replacing `neogeo.core=fba` to `neogeo.core=fbalpha`, causing all
subsequent passes (_i.e._ boots) to change `neogeo.core=fbalpha` again
to… `neogeo.core=fbalphalpha` 😱

This commit fixes the regexp to check that the matched core name is a
full name (_i.e._ it ends with space or end-of-line).