Validate `es_input.cfg` on startup
This script adds a new migration script, that checks on each boot if
`es_input.cfg` is valid (according to the RelaxNG definition introduced
in 63f771e), and if not, it retires the invalid file (with a `.invalid`
suffix for later manual inspection) and uses the default one (the one
from SHARE_INIT).

Since this migration script is the fourth one we introduce in less than
a month, and since the file we are adding them to is starting to get a
bit bloated, I moved them all to a new init script (`S14migrate`) that
gathers all migration scripts (some improvements could be made, they
have been added as comments in there).