Remove old web manager
This commit removes the Buildroot package for the legacy web manager. It
is not used on any board anymore.

It also adds a migration to bump webmanager from 1 to 2 in users
existing `recalbox.conf`.

Finally, it adds a warning in the logs when a user somehow manages to
try to start webmanager v1.

Signed-off-by: Michael Baudino <michael.baudino@alpine-lab.com>