fix(system): mount FAT32/ExFAT/NTFS devices with 755/644 rights
This commit explicitely mounts FAT32, ExFAT and NTFS external devices
with 755 as directory rights and 644 as file rights.

For NTFS and ExFAT, the default before that was 777 (for both files and
directories), which was an issue for dropbear (it's insecure, so
dropbear wouldn't use `~/.ssh/authorized_keys` with 777 rights, thus
disabling key-based SSH login).

More accurately, they use `umask=0` as default. See their man pages.

For FAT32, it was already the default, but the more explicit the better
(and the more futureproof: who knows if default won't change someday).