chore(ci): configure linters to run locally
This commit organizes linters in scripts that are included in the
codebase, rather than configured only for Gitlab CI.

This way, the Gitlab CI configuration is simplified (it simply calls
the scripts) and they can be run locally (either manually or triggered
by git hooks: I recommend linting commit messages in the `commit-msg`
hook and everything else in the `pre-commit` hook).

Those scripts are located in `test/lint` and a handy wrapper script has
been created at the root of the project for quick manual run:
`run_linters`.

This commit also adds a neat build badge on README ✨