docs: add useful comments in /boot/recalbox-boot.conf
This commit adds verbose comments in `recalbox.conf` that act as a
documentation for the supported variables and their different values.

Since this commit introduces comments in this file (lines starting with
a hash or a semi-colon), this commit also ensures that our parsing of
this file properly ignores those comments.