fix(os): make tic-80 downloader executable
Our shell script that downloads all content for the TIC80 emulator
didn't have the execution bit in its rights, thus could be executed by a
shell (_e.g._ `bash recalbox-tic80`) but not standalone (_e.g._
`./recalbox-tic80`).

This commit fixes this by simply adding the execution bit to the script.

This commit also adds some space and comments in the script for
readability.